feat: consolidate typing around file types
